+++ openbsc/src/bsc/osmo_bsc_api.c
@@ -44,7 +44,7 @@ static void bsc_dtap(struct gsm_subscriber_connection *conn, struct msgb *msg)
 {
 }
 
-static void bsc_assign_compl(struct gsm_subscriber_connection *conn, uint16_t rr_cause)
+static void bsc_assign_compl(struct gsm_subscriber_connection *conn, uint8_t rr_cause)
 {
 }
